event generator created, but need to learn how to sort it.

+++ src/main/java/seng302/Event.java
@@ -8,9 +8,9 @@ public class Event {
     private String eventName;
     public int eventOrder;
     public float distToPrevEvent;
-    public float nextHeading;
+    public int nextHeading;
 
-    public Event(String eventName, int eventOrder, float dist, float nextHeading) {
+    public Event(String eventName, int eventOrder, float dist, int nextHeading) {
         this.eventName = eventName;
         this.eventOrder = eventOrder;
         this.distToPrevEvent = dist;
